requirements.txt
appdirs==1.4.4
attrs==23.1.0
Automat==22.10.0
bagit==1.8.1
boto3==1.28.57
botocore==1.31.57
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
dateutils==0.6.12
filelock==3.12.4
fs==2.4.16
fs-s3fs==1.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
ocflcore==0.1.0
packaging==23.1
Pairtree==0.8.1
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
tldextract==3.6.0
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==1.26.16
w3lib==2.1.2
zope.interface==6.0

python-frontmatter~=1.1.0